+++ livechat/reports/api/v35.py
@@ -8,6 +8,8 @@ from livechat.utils.helpers import prepare_payload
 from livechat.utils.http_client import HttpClient
 from livechat.utils.structures import AccessToken
 
+# pylint: disable=unused-argument,too-many-arguments
+
 
 class ReportsApiV35(HttpClient):
     ''' Reports API client class in version 3.5. '''
@@ -47,7 +49,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                            which contains a server’s response to an HTTP request.
+                            which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -77,7 +79,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -107,7 +109,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -137,7 +139,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -167,7 +169,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -197,7 +199,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -227,7 +229,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                            which contains a server’s response to an HTTP request.
+                            which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -235,12 +237,13 @@ class ReportsApiV35(HttpClient):
                                  json=payload,
                                  headers=headers)
 
-    def surveys(self,
-                timezone: str = None,
-                filters: dict = None,
-                payload: dict = None,
-                headers: dict = None) -> httpx.Response:
-        ''' Returns the number of submitted chat surveys along with the count of specific answers.
+    def forms(self,
+              timezone: str = None,
+              filters: dict = None,
+              payload: dict = None,
+              headers: dict = None) -> httpx.Response:
+        ''' Returns the number of submitted chat forms along with the count of specific answers.
+
 
         Args:
             timezone (str): IANA Time Zone (e.g. America/Phoenix).
@@ -255,11 +258,11 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                            which contains a server’s response to an HTTP request.
+                            which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
-        return self.session.post(f'{self.api_url}/chats/surveys',
+        return self.session.post(f'{self.api_url}/chats/forms',
                                  json=payload,
                                  headers=headers)
 
@@ -285,7 +288,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                            which contains a server’s response to an HTTP request.
+                            which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -315,7 +318,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                            which contains a server’s response to an HTTP request.
+                            which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -347,7 +350,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -377,7 +380,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())
@@ -408,7 +411,7 @@ class ReportsApiV35(HttpClient):
 
         Returns:
             httpx.Response: The Response object from `httpx` library,
-                                which contains a server’s response to an HTTP request.
+                                which contains a server's response to an HTTP request.
         '''
         if payload is None:
             payload = prepare_payload(locals())